Create client options trait
